# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Roadcapture API
  description: Roadcapture API Description
  version: 0.1.0
servers:
- url: http://localhost:8080
...
paths:
...
  /users:
    get:
      tags:
      - users
      summary: 회원조회
      description: 회원조회
      operationId: '회원조회 - '
      parameters:
      - name: page
        in: query
        description: 조회할 페이지입니다. 0부터 시작합니다.
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: 한 페이지에 보여줄 사이즈 수입니다.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: 정렬 기준입니다.
        required: false
        schema:
          type: string
      - name: username
        in: query
        description: 조회할 사용자 이름입니다.
        required: false
        schema:
          type: string
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-804852034'
              examples:
                회원조회 - 성공:
                  value: "{\"content\":[{\"id\":1,\"username\":\"user1\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},{\"id\":2,\"username\":\"user2\"\
                    ,\"profileImageUrl\":\"https://www.test.com/test\"},{\"id\":3,\"\
                    username\":\"user3\",\"profileImageUrl\":\"https://www.test.com/test\"\
                    },{\"id\":4,\"username\":\"user4\",\"profileImageUrl\":\"https://www.test.com/test\"\
                    },{\"id\":5,\"username\":\"user5\",\"profileImageUrl\":\"https://www.test.com/test\"\
                    }],\"pageable\":{\"sort\":{\"sorted\":true,\"unsorted\":false,\"\
                    empty\":false},\"pageNumber\":0,\"pageSize\":10,\"offset\":0,\"\
                    paged\":true,\"unpaged\":false},\"last\":true,\"totalElements\"\
                    :5,\"totalPages\":1,\"number\":0,\"sort\":{\"sorted\":true,\"\
                    unsorted\":false,\"empty\":false},\"first\":true,\"numberOfElements\"\
                    :5,\"size\":10,\"empty\":false}"
                회원조회 - 유저 이름으로 검색 성공:
                  value: "{\"content\":[{\"id\":2,\"username\":\"user2\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"}],\"pageable\":{\"sort\":{\"sorted\"\
                    :true,\"unsorted\":false,\"empty\":false},\"pageNumber\":0,\"\
                    pageSize\":10,\"offset\":0,\"paged\":true,\"unpaged\":false},\"\
                    last\":true,\"totalElements\":1,\"totalPages\":1,\"number\":0,\"\
                    sort\":{\"sorted\":true,\"unsorted\":false,\"empty\":false},\"\
                    first\":true,\"numberOfElements\":1,\"size\":10,\"empty\":false}"
    post:
      tags:
      - users
      summary: 회원가입
      description: 회원가입
      operationId: '회원가입 - '
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/users1083751850'
            examples:
              회원가입 - 기가입 시 실패:
                value: "{\"email\":\"user1@gmail.com\",\"password\":\"abcd1234\",\"\
                  username\":\"test\",\"profileImageUrl\":null,\"provider\":null}"
              회원가입 - 닉네임 길이가 12보다 길면 실패:
                value: "{\"email\":\"test@gmail.com\",\"password\":\"abcd1234\",\"\
                  username\":\"testtesttesttest\",\"profileImageUrl\":null,\"provider\"\
                  :null}"
              회원가입 - 닉네임 길이가 2보다 짧으면 실패:
                value: "{\"email\":\"test@gmail.com\",\"password\":\"abcd1234\",\"\
                  username\":\"t\",\"profileImageUrl\":null,\"provider\":null}"
              회원가입 - 비밀번호 길이가 64보다 길면 실패:
                value: "{\"email\":\"test@gmail.com\",\"password\":\"abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789\"\
                  ,\"username\":\"test\",\"profileImageUrl\":null,\"provider\":null}"
              회원가입 - 비밀번호 길이가 8보다 짧으면 실패:
                value: "{\"email\":\"test@gmail.com\",\"password\":\"abcd123\",\"\
                  username\":\"test\",\"profileImageUrl\":null,\"provider\":null}"
              회원가입 - 비밀번호 형식 맞지 않으면 실패:
                value: "{\"email\":\"test@gmail.com\",\"password\":\"abcdefgh\",\"\
                  username\":\"test\",\"profileImageUrl\":null,\"provider\":null}"
              회원가입 - 성공:
                value: "{\"email\":\"test@gmail.com\",\"password\":\"abcd1234\",\"\
                  username\":\"test\",\"profileImageUrl\":null,\"provider\":null}"
              회원가입 - 이메일 형식 맞지 않으면 실패:
                value: "{\"email\":\"test\",\"password\":\"abcd1234\",\"username\"\
                  :\"test\",\"profileImageUrl\":null,\"provider\":null}"
      responses:
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-395173474'
              examples:
                회원가입 - 기가입 시 실패:
                  value: "{\"code\":\"BIZ-002\",\"status\":400,\"message\":\"이미 가입\
                    한 사용자입니다.\",\"errors\":[]}"
                회원가입 - 닉네임 길이가 12보다 길면 실패:
                  value: "{\"code\":\"CMM-001\",\"status\":400,\"message\":\"잘못된 입\
                    력입니다.\",\"errors\":[{\"field\":\"username\",\"value\":\"testtesttesttest\"\
                    ,\"reason\":\"size must be between 2 and 12\"}]}"
                회원가입 - 닉네임 길이가 2보다 짧으면 실패:
                  value: "{\"code\":\"CMM-001\",\"status\":400,\"message\":\"잘못된 입\
                    력입니다.\",\"errors\":[{\"field\":\"username\",\"value\":\"t\",\"\
                    reason\":\"size must be between 2 and 12\"}]}"
                회원가입 - 비밀번호 길이가 64보다 길면 실패:
                  value: "{\"code\":\"CMM-001\",\"status\":400,\"message\":\"잘못된 입\
                    력입니다.\",\"errors\":[{\"field\":\"password\",\"value\":\"abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789\"\
                    ,\"reason\":\"length must be between 8 and 64\"}]}"
                회원가입 - 비밀번호 길이가 8보다 짧으면 실패:
                  value: "{\"code\":\"CMM-001\",\"status\":400,\"message\":\"잘못된 입\
                    력입니다.\",\"errors\":[{\"field\":\"password\",\"value\":\"abcd123\"\
                    ,\"reason\":\"length must be between 8 and 64\"}]}"
                회원가입 - 비밀번호 형식 맞지 않으면 실패:
                  value: "{\"code\":\"CMM-001\",\"status\":400,\"message\":\"잘못된 입\
                    력입니다.\",\"errors\":[{\"field\":\"password\",\"value\":\"abcdefgh\"\
                    ,\"reason\":\"must match \\\"^(?=.*[A-Za-z])(?=.*\\\\d)[A-Za-z\\\
                    \\d]+$\\\"\"}]}"
                회원가입 - 이메일 형식 맞지 않으면 실패:
                  value: "{\"code\":\"CMM-001\",\"status\":400,\"message\":\"잘못된 입\
                    력입니다.\",\"errors\":[{\"field\":\"email\",\"value\":\"test\",\"\
                    reason\":\"must be a well-formed email address\"}]}"
        "201":
          description: "201"
    delete:
      tags:
      - users
      summary: 회원삭제
      description: 회원삭제
      operationId: '회원삭제 - '
      parameters:
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - users
      summary: 회원수정
      description: 회원수정
      operationId: 회원수정 - 성공
      parameters:
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/users-1546980116'
            examples:
              회원수정 - 성공:
                value: "{\"username\":\"updated101\",\"profileImageUrl\":\"https://test.com/updatedTest\"\
                  ,\"introduction\":\"안녕하세요. 저는 updated101입니다.\",\"address\":{\"addressName\"\
                  :\"경기 시흥시 정왕동 2121-3 경기과학기술대학\",\"roadAddressName\":\"경기 시흥시 경기과\
                  기대로 269 경기과학기술대학\",\"region1DepthName\":\"경기도\",\"region2DepthName\"\
                  :\"시흥시\",\"region3DepthName\":\"정왕동\",\"zoneNo\":\"15073\"}}"
      responses:
        "200":
          description: "200"
  /users/details:
    get:
      tags:
      - users
      summary: 회원상세조회
      description: 회원상세조회
      operationId: 회원상세조회 - 성공
      parameters:
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-details1674013768'
              examples:
                회원상세조회 - 성공:
                  value: "{\"id\":2,\"email\":\"user2@gmail.com\",\"username\":\"\
                    user2\",\"profileImageUrl\":\"https://www.test.com/test\",\"introduction\"\
                    :\"안녕하세요. 저는 user2입니다.\",\"address\":{\"addressName\":\"경기 시흥시\
                    \ 정왕동 2121-1 한국산업기술대학교\",\"roadAddressName\":\"경기 시흥시 산기대학로 237\
                    \ 한국산업기술대학교\",\"region1DepthName\":\"경기도\",\"region2DepthName\"\
                    :\"시흥시\",\"region3DepthName\":\"정왕동\",\"zoneNo\":\"15073\"}}"
  /users/token:
    post:
      tags:
      - users
      summary: 로그인
      description: 로그인
      operationId: 로그인 - 성공
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/users-token1713007146'
            examples:
              로그인 - 성공:
                value: "{\"email\":\"user1@gmail.com\",\"password\":\"abcd1234\",\"\
                  provider\":null}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-social-facebook-token1434215141'
              examples:
                로그인 - 성공:
                  value: "{\"grantType\":\"bearer\",\"accessToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.8bBaWeinsvgOpCZWgEp-Aw-3Uv8HujgoLcHkWWPCZmc\"\
                    ,\"refreshToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjMxMTIzNzE4OTF9.2jHKbwE_oLhCrc6kX9GcZ9hkkSnv17Yi96RCr8LHD8c\"\
                    ,\"accessTokenExpireDate\":3600000}"
  /users/{id}:
    get:
      tags:
      - users
      summary: 회원단건조회
      description: 회원단건조회
      operationId: '회원단건조회 - '
      parameters:
      - name: id
        in: path
        description: 조회할 사용자 아이디입니다.
        required: true
        schema:
          type: string
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-id1467703423'
              examples:
                회원단건조회 - 성공:
                  value: "{\"id\":2,\"username\":\"user2\",\"profileImageUrl\":\"\
                    https://www.test.com/test\",\"introduction\":\"안녕하세요. 저는 user2입\
                    니다.\"}"
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-social-facebook-883544378'
              examples:
                회원단건조회 - 회원 존재하지 않으면 실패:
                  value: "{\"code\":\"BIZ-003\",\"status\":400,\"message\":\"사용자가\
                    \ 존재하지 않습니다.\",\"errors\":[]}"
...
  /users/token/reissue:
    post:
      tags:
      - users
      summary: 토큰재발급
      description: 토큰재발급
      operationId: 토큰재발급 - 성공
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/users-token-reissue1815000917'
            examples:
              토큰재발급 - 성공:
                value: "{\"accessToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg\"\
                  ,\"refreshToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjMxMTIzNzE4OTF9.2jHKbwE_oLhCrc6kX9GcZ9hkkSnv17Yi96RCr8LHD8c\"\
                  }"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users-social-facebook-token1434215141'
              examples:
                토큰재발급 - 성공:
                  value: "{\"grantType\":\"bearer\",\"accessToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MywiZXhwIjoxNjQxMTQ2NTYzfQ.B4SlDH9e5FROPQfPtSkdGUPkO-5KkwrqZke0I3pGRBg\"\
                    ,\"refreshToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjMxMTIzNzE4OTF9.2jHKbwE_oLhCrc6kX9GcZ9hkkSnv17Yi96RCr8LHD8c\"\
                    ,\"accessTokenExpireDate\":3600000}"
...
components:
  schemas:
...
    users-id1467703423:
      type: object
      properties:
        id:
          type: number
          description: 사용자 아이디입니다.
        profileImageUrl:
          type: string
          description: 사용자 프로필 사진입니다.
        introduction:
          type: string
          description: 사용자 소개글입니다.
        username:
          type: string
          description: 사용자 이름입니다.
...
    users-1546980116:
      type: object
      properties:
        address:
          type: object
          properties:
            region2DepthName:
              type: string
              description: 시군구명입니다.
            zoneNo:
              type: string
              description: 우편번호입니다.
            region3DepthName:
              type: string
              description: 읍면동명입니다.
            roadAddressName:
              type: string
              description: 도로명주소입니다.
            region1DepthName:
              type: string
              description: 시구명입니다.
            addressName:
              type: string
              description: 지번주소입니다.
          description: 변경할 사용자 주소입니다.
        profileImageUrl:
          type: string
          description: 변경할 사용자 프로필 사진입니다. 주소 형식여야 합니다.
        introduction:
          type: string
          description: 변경할 사용자 소개입니다. 최대 200자 이하여야 합니다.
        username:
          type: string
          description: 변경할 사용자 이름입니다. 최소 2자 이상에서 최대 12자 이하여야 합니다.
    users-395173474:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드입니다.
        message:
          type: string
          description: 에러 메시지입니다.
        errors:
          type: array
          description: 필드 에러입니다. 필드 검증 시에만 존재합니다.
          items:
            type: object
            properties:
              reason:
                type: string
                description: 검증에 실패한 이유입니다.
              field:
                type: string
                description: 검증에 실패한 필드명입니다.
              value:
                type: string
                description: 검증에 실패한 필드값입니다.
        status:
          type: number
          description: 응답 코드입니다.
...
    users-details1674013768:
      type: object
      properties:
        address:
          type: object
          properties:
            region2DepthName:
              type: string
              description: 시군구명입니다.
            zoneNo:
              type: string
              description: 우편번호입니다.
            region3DepthName:
              type: string
              description: 읍면동명입니다.
            roadAddressName:
              type: string
              description: 도로명주소입니다.
            region1DepthName:
              type: string
              description: 시구명입니다.
            addressName:
              type: string
              description: 지번주소입니다.
          description: 사용자 주소입니다.
        id:
          type: number
          description: 사용자 아이디입니다.
        profileImageUrl:
          type: string
          description: 사용자 프로필 사진입니다.
        introduction:
          type: string
          description: 사용자 소개글입니다.
        email:
          type: string
          description: 사용자 이메일입니다.
        username:
          type: string
          description: 사용자 이름입니다.
    users-social-facebook-883544378:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드입니다.
        message:
          type: string
          description: 에러 메시지입니다.
        errors:
          type: array
          description: 필드 에러입니다. 필드 검증 시에만 존재합니다.
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        status:
          type: number
          description: 응답 코드입니다.
    users-token1713007146:
      type: object
      properties:
        password:
          type: string
          description: 사용자 비밀번호입니다.
        provider:
          type: string
          description: 사용자 정보 제공자입니다.
        email:
          type: string
          description: 사용자 이메일입니다.
    users1083751850:
      type: object
      properties:
        password:
          type: string
          description: 사용자 비밀번호입니다. 영문 숫자 조합 최소 8자 이상에서 최대 64자 이하여야 합니다.
        provider:
          type: string
          description: 사용자 정보 제공자입니다.
        profileImageUrl:
          type: string
          description: 사용자 프로필 이미지입니다.
        email:
          type: string
          description: 사용자 이메일입니다. 이메일 형식이어야 합니다.
        username:
          type: string
          description: 사용자 이름입니다. 최소 2자 이상에서 최대 12자 이하여야 합니다.
    users-804852034:
      type: object
      properties:
        size:
          type: number
          description: 한 페이지에 보여줄 사이즈 수입니다.
        last:
          type: boolean
          description: 마지막 페이지 여부입니다.
        totalPages:
          type: number
          description: 전체 페이지 수입니다.
        first:
          type: boolean
          description: 첫 페이지 여부입니다.
        content:
          type: array
          description: 페이지 요소 리스트입니다.
          items:
            type: object
            properties:
              id:
                type: number
                description: 사용자 아이디입니다.
              profileImageUrl:
                type: string
                description: 사용자 프로필 사진입니다.
              username:
                type: string
                description: 사용자 이름입니다.
        empty:
          type: boolean
          description: 리스트가 비어 있는지 여부입니다.
        totalElements:
          type: number
          description: 전체 요소 수입니다.
    users-social-facebook-token1434215141:
      type: object
      properties:
        accessTokenExpireDate:
          type: number
          description: jwt 입니다. api 요청 시 사용됩니다.
        accessToken:
          type: string
          description: 액세스 토큰입니다. api 요청 시 사용됩니다.
        grantType:
          type: string
          description: 승인 타입입니다.
        refreshToken:
          type: string
          description: 리프레시 토큰입니다. 토큰 재발급에 사용됩니다.
    users-token-reissue1815000917:
      type: object
      properties:
        accessToken:
          type: string
          description: 액세스 토큰입니다.
        refreshToken:
          type: string
          description: 리프레시 토큰입니다.
...